src/main/java/com/deerhunter/topic/Topic128.java
package com.deerhunter.topic;

import java.util.HashSet;
import java.util.Set;

/**
 * 给定一个未排序的整数数组，找出最长连续序列的长度。
 * <p>
 * 要求算法的时间复杂度为 O(n)。
 * <p>
 * 示例:
 * <p>
 * 输入: [100, 4, 200, 1, 3, 2]
 * 输出: 4
 * 解释: 最长连续序列是 [1, 2, 3, 4]。它的长度为 4。
 * <p>
 * 来源：力扣（LeetCode）
 * 链接：https://leetcode-cn.com/problems/longest-consecutive-sequence
 * 著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
 * <p>
 * Copyright (c) 20120 deerhunter0837@gmail.com All Rights Reserved.
 *
 * @author xuejunc
 * @createTime 2020/9/20 17:22
 */
public class Topic128 {
    /**
     * 参考官方题解
     */
    public static class Solution {
        public int longestConsecutive(int[] nums) {
            Set<Integer> set = new HashSet<>();
            for (int i = 0; i < nums.length; i++) {
                set.add(nums[i]);
            }
            int longestStreak = 0;
            for (Integer num : set) {
                if (!set.contains(num - 1)) {
                    int start = num;
                    int streak = 1;
                    while (set.contains(++start)) {
                        streak++;
                    }
                    longestStreak = Math.max(longestStreak, streak);
                }
            }
            return longestStreak;
        }

    }
}
